+++ src/IECoreRenderMan/MaterialCache.h
@@ -65,7 +65,7 @@ class MaterialCache
 		// Can be called concurrently with other calls to `get()`
 		ConstMaterialPtr getMaterial( const IECoreScene::ShaderNetwork *network );
 		ConstDisplacementPtr getDisplacement( const IECoreScene::ShaderNetwork *network );
-		ConstLightShaderPtr getLightShader( const IECoreScene::ShaderNetwork *network, const IECoreScene::ShaderNetwork *lightFilter );
+		ConstLightShaderPtr getLightShader( const IECoreScene::ShaderNetwork *network, const IECoreScene::ShaderNetwork *lightFilter, RtUString shadowSubset );
 
 		// Must not be called concurrently with anything.
 		void clearUnused();